requirements.txt
#
django
python-dotenv
django-cors-headers
djangorestframework-simplejwt
django-ckeditor
# psycopg2
psycopg2-binary
django-extensions
getmac
django-taggit
Pillow
django-currentuser
user-agents
django-user-agents
unidecode
requests
gunicorn
django-protected-media
djangorestframework-api-key
boto3
django-storages
cryptography

#
langchain-cli
langchain-experimental
langchain-openai
langchain-community
tiktoken
psycopg
pgvector


#
aiohttp
aiosignal
annotated-types
antlr4-python3-runtime
anyio
async-timeout
asyncpg
attrs
backoff
beautifulsoup4
cachetools
certifi
cffi
chardet
charset-normalizer
click
colorama
coloredlogs
contourpy
cryptography
cycler
dataclasses-json
deepdiff
Deprecated
distro
effdet
emoji
et-xmlfile
fastapi
filelock
filetype
flatbuffers
fonttools
frozenlist
fsspec
gitdb
GitPython
google-api-core
google-auth
google-cloud-vision
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
httpcore
httptools
httpx
huggingface-hub
humanfriendly
idna
intel-openmp
iopath
Jinja2
joblib
jsonpatch
jsonpath-python
jsonpointer
jsonschema
jsonschema-specifications
kiwisolver
langchain
langchain-cli
langchain-community
langchain-core
langchain-experimental
langchain-openai
langchain-text-splitters
langdetect
langserve
langsmith
layoutparser
libcst
lxml
Markdown
markdown-it-py
MarkupSafe
marshmallow
matplotlib
mdurl
mkl
mpmath
multidict
mypy-extensions
nest-asyncio
networkx
nltk
numpy
olefile
omegaconf
onnx
onnxruntime
openai
opencv-python
openpyxl
ordered-set
orjson
packaging
pandas
pdf2image
pdfminer.six
pdfplumber
pgvector
pikepdf
pillow
pillow_heif
pip
portalocker
proto-plus
protobuf
psutil
psycopg
psycopg2-binary
pyasn1
pyasn1_modules
pycocotools
pycparser
pydantic
pydantic_core
Pygments
pypandoc
pyparsing
pypdf
pypdfium2
pyproject-toml
pyreadline3
pytesseract
python-dateutil
python-docx
python-dotenv
python-iso639
python-magic
python-multipart
python-oxmsg
python-pptx
pytz
pywin32
PyYAML
rapidfuzz
referencing
regex
requests
requests-toolbelt
rich
rpds-py
rsa
safetensors
scipy
setuptools
shellingham
six
smmap
sniffio
soupsieve
SQLAlchemy
sse-starlette
starlette
sympy
tabulate
tbb
tenacity
tiktoken
timm
tokenizers
toml
tomlkit
torch
torchvision
tqdm
transformers
typer
typing_extensions
typing-inspect
tzdata
unstructured
unstructured-client
unstructured-inference
unstructured.pytesseract
urllib3
uvicorn
watchfiles
websockets
wheel
wrapt
xlrd
XlsxWriter
yarl